card_app/requirements.txt
flask==3.0.3
pillow==10.4.0
numpy==1.24.4
opencv-python==4.10.0.84
torch==2.1.2
torchvision==0.16.2
